backend/requirements.txt
django
djangorestframework
google-generativeai
channels
channels-redis
daphne
langchain
langchain-google-genai
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
whitenoise
xhtml2pdf
django-cors-headers
pymongo==4.6.2
dnspython==2.6.1
markdown
cloudinary
PyPDF2
reportlab